% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensitivity.R
\name{alignment}
\alias{alignment}
\title{Alignment confounding function (ATE)}
\usage{
alignment(alpha, p, treatment)
}
\arguments{
\item{alpha}{scalar or vector}

\item{p}{propensity score vector}

\item{treatment}{treatment vector}
}
\value{
adjustment vector
}
\description{
Alignment confounding function (ATE)
}
\examples{
\dontrun{
# Basic usage
# alignment(...)
}
}
\seealso{
\code{\link{RCausalML-package}}
}
