% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensitivity.R
\name{alignment_att}
\alias{alignment_att}
\title{Alignment confounding function for ATT}
\usage{
alignment_att(alpha, p, treatment)
}
\arguments{
\item{alpha}{scalar or vector}

\item{p}{propensity score vector}

\item{treatment}{treatment vector}
}
\value{
adjustment vector
}
\description{
Alignment confounding function for ATT
}
\examples{
\dontrun{
# Basic usage
# alignment_att(...)
}
}
\seealso{
\code{\link{RCausalML-package}}
}
