% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/causalBoosted_IV.R
\name{ATE_error}
\alias{ATE_error}
\title{Absolute Error in Average Treatment Effect}
\usage{
ATE_error(tau_true, tau_hat)
}
\arguments{
\item{tau_true}{Numeric vector of true individual treatment effects.}

\item{tau_hat}{Numeric vector of estimated individual treatment effects.}
}
\value{
Numeric scalar.
}
\description{
\eqn{|E[\tau(x)] - E[\hat\tau(x)]|}.
}
\examples{
ATE_error(c(1, 2, 3), c(1.1, 1.9, 3.2))
}
\seealso{
\code{\link{RCausalML-package}}
}
