% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drawing.R
\name{bfs_layout}
\alias{bfs_layout}
\title{BFS layout: layers by breadth-first search from a start node}
\usage{
bfs_layout(
  G,
  start,
  align = c("vertical", "horizontal"),
  scale = 1,
  center = NULL,
  as_list = FALSE
)
}
\arguments{
\item{G}{igraph object or adjacency matrix.}

\item{start}{Start node (vertex id or name).}

\item{align}{"vertical" or "horizontal".}

\item{scale}{Scale factor.}

\item{center}{Center coordinates.}
}
\value{
Matrix of positions (n x 2) or list if \code{as_list=TRUE}.
}
\description{
BFS layout: layers by breadth-first search from a start node
}
\examples{
\dontrun{
# Basic usage
# bfs_layout(...)
}
}
\seealso{
\code{\link{RCausalML-package}}
}
