% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/causalDeepNet.R
\name{causal_structure_ml_model_descriptions}
\alias{causal_structure_ml_model_descriptions}
\title{Short descriptions of causal structure learning methods}
\usage{
causal_structure_ml_model_descriptions()
}
\value{
Named character vector.
}
\description{
Returns one-line summaries of each \code{\link{causalStructureML}} \code{method}.
}
\examples{
\dontrun{
# Basic usage
# causal_structure_ml_model_descriptions(...)
}
}
\seealso{
\code{\link{RCausalML-package}}
}
