% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/causalDeepNet.R
\name{CausalDiscrepancyVAE}
\alias{CausalDiscrepancyVAE}
\title{Alias for \code{causal_discrepancy_vae}}
\usage{
CausalDiscrepancyVAE(X, treatment, y, ...)
}
\arguments{
\item{X}{Covariate matrix or data.frame.}

\item{treatment}{Binary treatment vector (0/1).}

\item{y}{Outcome vector.}

\item{...}{Ignored.}
}
\value{
Object of class \code{causal_discrepancy_vae}.
}
\description{
Alias for \code{causal_discrepancy_vae}
}
\examples{
\dontrun{
# Basic usage
# CausalDiscrepancyVAE(...)
}
}
\seealso{
\code{\link{RCausalML-package}}
}
