% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/causalDeepNet.R
\name{CausalEGM}
\alias{CausalEGM}
\title{Alias for \code{causal_egm}}
\usage{
CausalEGM(X, treatment, y, ...)
}
\arguments{
\item{X}{Covariate matrix or data.frame.}

\item{treatment}{Binary treatment vector (0/1).}

\item{y}{Outcome vector.}

\item{...}{Ignored.}
}
\value{
Object of class \code{causal_egm}.
}
\description{
Alias for \code{causal_egm}
}
\examples{
\dontrun{
# Basic usage
# CausalEGM(...)
}
}
\seealso{
\code{\link{RCausalML-package}}
}
