% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drawing.R
\name{circular_layout}
\alias{circular_layout}
\title{Circular layout: place nodes on a circle}
\usage{
circular_layout(G, scale = 1, center = NULL, dim = 2L, as_list = FALSE)
}
\arguments{
\item{G}{igraph object or adjacency matrix.}

\item{scale}{Scale factor for radius.}

\item{center}{Center coordinates.}

\item{dim}{Dimension (2 only for circle).}
}
\value{
Matrix of positions (n x 2) or list if \code{as_list=TRUE}.
}
\description{
Circular layout: place nodes on a circle
}
\examples{
\dontrun{
# Basic usage
# circular_layout(...)
}
}
\seealso{
\code{\link{RCausalML-package}}
}
