% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{classification_metrics}
\alias{classification_metrics}
\title{Classification metrics (AUC proxy, accuracy) by treatment group}
\usage{
classification_metrics(y, yhat, w)
}
\arguments{
\item{y}{observed binary outcome}

\item{yhat}{predicted probability}

\item{w}{treatment indicator}
}
\value{
Object returned by \code{classification_metrics}.
}
\description{
Classification metrics (AUC proxy, accuracy) by treatment group
}
\examples{
\dontrun{
# Basic usage
# classification_metrics(...)
}
}
\seealso{
\code{\link{RCausalML-package}}
}
