% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DMLearner.R
\name{coef.LinearDML}
\alias{coef.LinearDML}
\alias{coef.SparseLinearDML}
\title{Coefficients of the linear CATE model (LinearDML or SparseLinearDML)}
\usage{
\method{coef}{LinearDML}(object, ...)

\method{coef}{SparseLinearDML}(object, ...)
}
\value{
Object returned by \code{coef.LinearDML}.

Object returned by \code{coef.SparseLinearDML}.
}
\description{
Coefficients of the linear CATE model (LinearDML or SparseLinearDML)
}
\examples{
\dontrun{
# Basic usage
# coef.LinearDML(...)
}
\dontrun{
# Basic usage
# coef.SparseLinearDML(...)
}
}
\seealso{
\code{\link{RCausalML-package}}

\code{\link{RCausalML-package}}
}
