% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimize.R
\name{counterfactual_value_estimator}
\alias{counterfactual_value_estimator}
\title{Counterfactual value estimator: E[(v - cc_w) * Y_w - ic_w]}
\usage{
counterfactual_value_estimator(
  treatment,
  control_name = "control",
  treatment_names,
  y_proba,
  cate = NULL,
  value = 1,
  conversion_cost = 0,
  impression_cost = 0
)
}
\arguments{
\item{treatment}{vector of assigned treatment (group names or 0/1)}

\item{control_name}{name of control}

\item{treatment_names}{names of treatment arms}

\item{y_proba}{predicted probability of conversion (or outcome) under each treatment; matrix n x n_treatments or list}

\item{cate}{CATE estimates (optional) for value}

\item{value}{conversion value per unit}

\item{conversion_cost}{cost per conversion by treatment}

\item{impression_cost}{cost per impression by treatment}
}
\value{
object with predict_best method
}
\description{
Counterfactual value estimator: E[(v - cc_w) * Y_w - ic_w]
}
\examples{
\dontrun{
# Basic usage
# counterfactual_value_estimator(...)
}
}
\seealso{
\code{\link{RCausalML-package}}
}
