% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notears.R
\name{demo_linear}
\alias{demo_linear}
\title{Demo: linear NOTEARS on synthetic data}
\usage{
demo_linear()
}
\value{
Object returned by \code{demo_linear}.
}
\description{
Demo: linear NOTEARS on synthetic data
}
\examples{
\dontrun{
# Basic usage
# demo_linear(...)
}
}
\seealso{
\code{\link{RCausalML-package}}
}
