% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notears.R
\name{demo_nonlinear_mlp}
\alias{demo_nonlinear_mlp}
\title{Demo: nonlinear NOTEARS (MLP) on synthetic data}
\usage{
demo_nonlinear_mlp()
}
\value{
Object returned by \code{demo_nonlinear_mlp}.
}
\description{
Demo: nonlinear NOTEARS (MLP) on synthetic data
}
\examples{
\dontrun{
# Basic usage
# demo_nonlinear_mlp(...)
}
}
\seealso{
\code{\link{RCausalML-package}}
}
