% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notears.R
\name{demo_nonlinear_sobolev}
\alias{demo_nonlinear_sobolev}
\title{Demo: nonlinear NOTEARS (Sobolev) on random data}
\usage{
demo_nonlinear_sobolev()
}
\value{
Object returned by \code{demo_nonlinear_sobolev}.
}
\description{
Demo: nonlinear NOTEARS (Sobolev) on random data
}
\examples{
\dontrun{
# Basic usage
# demo_nonlinear_sobolev(...)
}
}
\seealso{
\code{\link{RCausalML-package}}
}
