% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/causalForest.R
\name{.compute_oob_variance}
\alias{.compute_oob_variance}
\title{OOB variance estimate (used when newdata = NULL)}
\usage{
.compute_oob_variance(object)
}
\description{
OOB variance estimate (used when newdata = NULL)
}
