% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/causalForest.R
\name{.compute_variance_ij}
\alias{.compute_variance_ij}
\title{Variance estimation for new-data predictions via grouped IJ}
\usage{
.compute_variance_ij(pred_matrix, ci.group.size)
}
\description{
Variance estimation for new-data predictions via grouped IJ
}
