% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/causalXGBoost.R
\name{.cxgb_apply_imputation}
\alias{.cxgb_apply_imputation}
\title{Apply a pre-built imputation table to a matrix}
\usage{
.cxgb_apply_imputation(X, table)
}
\description{
Apply a pre-built imputation table to a matrix
}
\keyword{internal}
