% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/causalXGBoost.R
\name{.cxgb_build_imputation}
\alias{.cxgb_build_imputation}
\title{Median imputation table (per column)}
\usage{
.cxgb_build_imputation(X)
}
\description{
Median imputation table (per column)
}
\keyword{internal}
