% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/causalForest.R
\name{.estimate_propensity}
\alias{.estimate_propensity}
\title{Compute propensity scores via logistic regression or provided values}
\usage{
.estimate_propensity(X, W)
}
\description{
Compute propensity scores via logistic regression or provided values
}
