% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/causalForest.R
\name{.predict_tree}
\alias{.predict_tree}
\title{Route a single observation through a tree, return leaf node}
\usage{
.predict_tree(tree, x_row)
}
\description{
Route a single observation through a tree, return leaf node
}
