% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DMLearner.R
\name{doubleml_data_from_matrix}
\alias{doubleml_data_from_matrix}
\title{Build \code{DoubleMLData} from matrices (\code{X}, \code{y}, \code{d}, optional \code{z}, etc.)}
\usage{
doubleml_data_from_matrix(
  X = NULL,
  y,
  d,
  z = NULL,
  s = NULL,
  cluster_vars = NULL,
  data_class = "DoubleMLData",
  use_other_treat_as_covariate = TRUE
)
}
\arguments{
\item{X, y, d, z, s, cluster_vars, data_class, use_other_treat_as_covariate}{As in \pkg{DoubleML}.}
}
\value{
\code{DoubleMLData}, \code{DoubleMLClusterData}, or \code{data.table} per \code{data_class};
  \code{NULL} if \pkg{DoubleML} is not installed.
}
\description{
Wrapper around \code{DoubleML::double_ml_data_from_matrix()}.
}
\examples{
\dontrun{
# Basic usage
# doubleml_data_from_matrix(...)
}
}
\references{
DoubleML: \url{https://docs.doubleml.org}
}
\seealso{
\code{\link{RCausalML-package}}
}
