% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DMLearner.R
\name{doubleml_fetch_401k}
\alias{doubleml_fetch_401k}
\title{401(k) data (\pkg{DoubleML} \code{fetch_401k})}
\usage{
doubleml_fetch_401k(...)
}
\arguments{
\item{...}{Arguments passed to \code{DoubleML::fetch_401k}.}
}
\value{
As in \pkg{DoubleML}; \code{NULL} if \pkg{DoubleML} is not installed.
}
\description{
401(k) data (\pkg{DoubleML} \code{fetch_401k})
}
\examples{
\dontrun{
# Basic usage
# doubleml_fetch_401k(...)
}
}
\seealso{
\code{\link{RCausalML-package}}
}
