% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DMLearner.R
\name{doubleml_fetch_bonus}
\alias{doubleml_fetch_bonus}
\title{Pennsylvania bonus experiment data (\pkg{DoubleML} \code{fetch_bonus})}
\usage{
doubleml_fetch_bonus(...)
}
\arguments{
\item{...}{Arguments passed to \code{DoubleML::fetch_bonus}.}
}
\value{
As in \pkg{DoubleML}; \code{NULL} if not installed.
}
\description{
Pennsylvania bonus experiment data (\pkg{DoubleML} \code{fetch_bonus})
}
\examples{
\dontrun{
# Basic usage
# doubleml_fetch_bonus(...)
}
}
\seealso{
\code{\link{RCausalML-package}}
}
