% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drawing.R
\name{draw_networkx_edges}
\alias{draw_networkx_edges}
\title{Draw network edges}
\usage{
draw_networkx_edges(
  G,
  pos,
  edge.color = "gray40",
  edge.width = 1,
  arrow.size = 0.2
)
}
\arguments{
\item{G}{igraph object.}

\item{pos}{Matrix or list of node positions.}

\item{edge.color}{Color(s) for edges.}

\item{edge.width}{Width(s).}

\item{arrow.size}{Arrow head length for directed edges; 0 for no arrows.}
}
\value{
Object returned by \code{draw_networkx_edges}.
}
\description{
Draw network edges
}
\examples{
\dontrun{
# Basic usage
# draw_networkx_edges(...)
}
}
\seealso{
\code{\link{RCausalML-package}}
}
