% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drawing.R
\name{draw_networkx_labels}
\alias{draw_networkx_labels}
\title{Draw node labels}
\usage{
draw_networkx_labels(G, pos, labels = NULL, cex = 0.8, col = "black")
}
\arguments{
\item{G}{igraph object.}

\item{pos}{Matrix or list of node positions.}

\item{labels}{Character vector of labels (same order as nodes in G).}

\item{cex}{Font size.}

\item{col}{Label color.}
}
\value{
Object returned by \code{draw_networkx_labels}.
}
\description{
Draw node labels
}
\examples{
\dontrun{
# Basic usage
# draw_networkx_labels(...)
}
}
\seealso{
\code{\link{RCausalML-package}}
}
