% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DMLearner.R
\name{effect.DynamicDMLearner}
\alias{effect.DynamicDMLearner}
\title{Effect of treatment change (DynamicDMLearner)}
\usage{
\method{effect}{DynamicDMLearner}(object, newdata = NULL, T0 = 0, T1 = 1, ...)
}
\arguments{
\item{T0}{baseline treatment (vector or matrix)}

\item{T1}{alternative treatment (same shape as T0)}
}
\value{
Object returned by \code{effect.DynamicDMLearner}.
}
\description{
Effect of treatment change (DynamicDMLearner)
}
\examples{
\dontrun{
# Basic usage
# effect.DynamicDMLearner(...)
}
}
\seealso{
\code{\link{RCausalML-package}}
}
