% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DMLearner.R
\name{effect}
\alias{effect}
\title{Effect of treatment change (generic)}
\usage{
effect(object, ...)
}
\arguments{
\item{object}{fitted model (e.g. DynamicDMLearner)}

\item{...}{passed to methods}
}
\value{
Object returned by \code{effect}.
}
\description{
Effect of treatment change (generic)
}
\examples{
\dontrun{
# Basic usage
# effect(...)
}
}
\seealso{
\code{\link{RCausalML-package}}
}
