% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_learners.R
\name{estimate_ate.default}
\alias{estimate_ate.default}
\title{Default ATE estimation dispatcher}
\usage{
\method{estimate_ate}{default}(
  obj,
  X,
  treatment,
  y,
  p = NULL,
  return_ci = TRUE,
  bootstrap_ci = FALSE,
  n_bootstraps = 1000L,
  bootstrap_size = 10000L,
  n_cores = 1L,
  pretrain = FALSE,
  learner = "ranger",
  ...
)
}
\value{
Object returned by \code{estimate_ate.default}.
}
\description{
Default ATE estimation dispatcher
}
\examples{
\dontrun{
# Basic usage
# estimate_ate.default(...)
}
}
\seealso{
\code{\link{RCausalML-package}}
}
