% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DMLearner.R
\name{estimate_ate.DMLearner}
\alias{estimate_ate.DMLearner}
\title{Estimate ATE for DMLearner}
\usage{
\method{estimate_ate}{DMLearner}(
  obj,
  X,
  treatment,
  y,
  p = NULL,
  return_ci = TRUE,
  bootstrap_ci = FALSE,
  n_bootstraps = 1000L,
  bootstrap_size = 10000L,
  n_cores = 1L,
  seed = NULL,
  pretrain = FALSE,
  ...
)
}
\value{
Object returned by \code{estimate_ate.DMLearner}.
}
\description{
Estimate ATE for DMLearner
}
\examples{
\dontrun{
# Basic usage
# estimate_ate.DMLearner(...)
}
}
\seealso{
\code{\link{RCausalML-package}}
}
