% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_learners.R
\name{estimate_ate.DomainAdaptationLearner}
\alias{estimate_ate.DomainAdaptationLearner}
\title{Estimate ATE for DomainAdaptationLearner}
\usage{
\method{estimate_ate}{DomainAdaptationLearner}(
  obj,
  X,
  treatment,
  y,
  p = NULL,
  return_ci = TRUE,
  bootstrap_ci = FALSE,
  n_bootstraps = 1000L,
  bootstrap_size = 10000L,
  n_cores = 1L,
  pretrain = FALSE,
  ...
)
}
\value{
Object returned by \code{estimate_ate.DomainAdaptationLearner}.
}
\description{
Estimate ATE for DomainAdaptationLearner
}
\examples{
\dontrun{
# Basic usage
# estimate_ate.DomainAdaptationLearner(...)
}
}
\seealso{
\code{\link{RCausalML-package}}
}
