% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/policy_learner.R
\name{feature_importances.DRPolicyForest}
\alias{feature_importances.DRPolicyForest}
\title{Feature importances (average over trees)}
\usage{
\method{feature_importances}{DRPolicyForest}(object, ...)
}
\arguments{
\item{object}{fitted \code{DRPolicyForest}.}

\item{...}{unused.}
}
\value{
Object returned by \code{feature_importances.DRPolicyForest}.
}
\description{
Feature importances (average over trees)
}
\examples{
\dontrun{
# Basic usage
# feature_importances.DRPolicyForest(...)
}
}
\seealso{
\code{\link{RCausalML-package}}
}
