% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/policy_learner.R
\name{feature_importances.DRPolicyTree}
\alias{feature_importances.DRPolicyTree}
\title{Feature importances from policy tree (variable importance from rpart)}
\usage{
\method{feature_importances}{DRPolicyTree}(object, ...)
}
\arguments{
\item{object}{fitted \code{DRPolicyTree}.}

\item{...}{unused.}
}
\value{
Named numeric vector of importance (sum to 1 if available).
}
\description{
Feature importances from policy tree (variable importance from rpart)
}
\examples{
\dontrun{
# Basic usage
# feature_importances.DRPolicyTree(...)
}
}
\seealso{
\code{\link{RCausalML-package}}
}
