% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/policy_learner.R
\name{feature_importances}
\alias{feature_importances}
\title{Feature importances for policy / tree models}
\usage{
feature_importances(object, ...)
}
\arguments{
\item{object}{fitted model.}

\item{...}{passed to methods.}
}
\value{
Object returned by \code{feature_importances}.
}
\description{
Feature importances for policy / tree models
}
\examples{
\dontrun{
# Basic usage
# feature_importances(...)
}
}
\seealso{
\code{\link{RCausalML-package}}
}
