% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DMLearner.R
\name{fit.DMLIVLearner}
\alias{fit.DMLIVLearner}
\title{Fit DMLIVLearner (native R)}
\usage{
\method{fit}{DMLIVLearner}(obj, Y, T, Z, X = NULL, W = NULL, ...)
}
\value{
Object returned by \code{fit.DMLIVLearner}.
}
\description{
Fit DMLIVLearner (native R)
}
\examples{
\dontrun{
# Basic usage
# fit.DMLIVLearner(...)
}
}
\seealso{
\code{\link{RCausalML-package}}
}
