% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_learners.R
\name{fit.DomainAdaptationLearner}
\alias{fit.DomainAdaptationLearner}
\title{Fit a DomainAdaptationLearner model}
\usage{
\method{fit}{DomainAdaptationLearner}(obj, X, treatment, y, p = NULL, ...)
}
\value{
Object returned by \code{fit.DomainAdaptationLearner}.
}
\description{
Fit a DomainAdaptationLearner model
}
\examples{
\dontrun{
# Basic usage
# fit.DomainAdaptationLearner(...)
}
}
\seealso{
\code{\link{RCausalML-package}}
}
