% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_learners.R
\name{fit.DRLearner}
\alias{fit.DRLearner}
\title{Fit DR-Learner}
\usage{
\method{fit}{DRLearner}(obj, X, treatment, y, p = NULL, ...)
}
\value{
Object returned by \code{fit.DRLearner}.
}
\description{
Fit DR-Learner
}
\examples{
\dontrun{
# Basic usage
# fit.DRLearner(...)
}
}
\seealso{
\code{\link{RCausalML-package}}
}
