% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DMLearner.R
\name{fit.OrthoIVLearner}
\alias{fit.OrthoIVLearner}
\title{Fit OrthoIVLearner (native R)}
\usage{
\method{fit}{OrthoIVLearner}(obj, Y, T, Z, X = NULL, W = NULL, ...)
}
\arguments{
\item{Y}{outcome}

\item{T}{treatment}

\item{Z}{instrument}

\item{X, W}{optional covariates and controls}
}
\value{
Object returned by \code{fit.OrthoIVLearner}.
}
\description{
Fit OrthoIVLearner (native R)
}
\examples{
\dontrun{
# Basic usage
# fit.OrthoIVLearner(...)
}
}
\seealso{
\code{\link{RCausalML-package}}
}
