% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_learners.R
\name{fit}
\alias{fit}
\title{Generic fit for meta-learners}
\usage{
fit(obj, X, treatment, y, p = NULL, ...)
}
\value{
Object returned by \code{fit}.
}
\description{
Generic fit for meta-learners
}
\examples{
\dontrun{
# Basic usage
# fit(...)
}
}
\seealso{
\code{\link{RCausalML-package}}
}
