% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_learners.R
\name{fit.SLearner}
\alias{fit.SLearner}
\title{Fit S-Learner}
\usage{
\method{fit}{SLearner}(obj, X, treatment, y, p = NULL, ...)
}
\arguments{
\item{obj}{SLearner object}

\item{X}{covariate matrix}

\item{treatment}{treatment vector (0/1 or control/treatment)}

\item{y}{outcome vector}

\item{p}{optional propensity (unused in S-learner)}
}
\value{
fitted SLearner
}
\description{
Fit S-Learner
}
\examples{
\dontrun{
# Basic usage
# fit.SLearner(...)
}
}
\seealso{
\code{\link{RCausalML-package}}
}
