% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_learners.R
\name{fit.TMLELearner}
\alias{fit.TMLELearner}
\title{Fit TMLE learner (stores outcome model; TMLE step done in estimate_ate)}
\usage{
\method{fit}{TMLELearner}(obj, X, treatment, y, p = NULL, ...)
}
\value{
Object returned by \code{fit.TMLELearner}.
}
\description{
Fit TMLE learner (stores outcome model; TMLE step done in estimate_ate)
}
\examples{
\dontrun{
# Basic usage
# fit.TMLELearner(...)
}
}
\seealso{
\code{\link{RCausalML-package}}
}
