% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DMLearner.R
\name{fit_predict.DMLearner}
\alias{fit_predict.DMLearner}
\title{Fit and predict (DMLearner)}
\usage{
\method{fit_predict}{DMLearner}(
  obj,
  X,
  treatment,
  y,
  p = NULL,
  return_components = FALSE,
  seed = NULL,
  ...
)
}
\value{
Object returned by \code{fit_predict.DMLearner}.
}
\description{
Fit and predict (DMLearner)
}
\examples{
\dontrun{
# Basic usage
# fit_predict.DMLearner(...)
}
}
\seealso{
\code{\link{RCausalML-package}}
}
