% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_learners.R
\name{fit_predict.DomainAdaptationLearner}
\alias{fit_predict.DomainAdaptationLearner}
\title{Fit and predict with DomainAdaptationLearner}
\usage{
\method{fit_predict}{DomainAdaptationLearner}(
  obj,
  X,
  treatment,
  y,
  p = NULL,
  return_ci = FALSE,
  n_bootstraps = 1000L,
  bootstrap_size = 10000L,
  n_cores = 1L,
  return_components = FALSE,
  verbose = TRUE,
  ...
)
}
\value{
Object returned by \code{fit_predict.DomainAdaptationLearner}.
}
\description{
Fit and predict with DomainAdaptationLearner
}
\examples{
\dontrun{
# Basic usage
# fit_predict.DomainAdaptationLearner(...)
}
}
\seealso{
\code{\link{RCausalML-package}}
}
