% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_learners.R
\name{fit_predict}
\alias{fit_predict}
\title{Generic fit_predict for meta-learners}
\usage{
fit_predict(obj, X, treatment, y, p = NULL, ...)
}
\value{
Object returned by \code{fit_predict}.
}
\description{
Generic fit_predict for meta-learners
}
\examples{
\dontrun{
# Basic usage
# fit_predict(...)
}
}
\seealso{
\code{\link{RCausalML-package}}
}
