% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drawing.R
\name{fruchterman_reingold_layout}
\alias{fruchterman_reingold_layout}
\title{Alias for spring_layout (NetworkX name)}
\usage{
fruchterman_reingold_layout(
  G,
  seed = NULL,
  iterations = 50,
  weight = "weight",
  scale = 1,
  center = NULL,
  as_list = FALSE
)
}
\value{
Object returned by \code{function_name}.
}
\description{
Alias for spring_layout (NetworkX name)
}
\examples{
\dontrun{
# Basic usage
# function_name(...)
}
}
\seealso{
\code{\link{RCausalML-package}}
}
