% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{gain_curve}
\alias{gain_curve}
\title{Gain curve: cumulative sum of true tau (normalized to 1) when ordered by pred}
\usage{
gain_curve(tau, pred)
}
\arguments{
\item{tau}{true treatment effect vector}

\item{pred}{predicted CATE vector}
}
\value{
normalized cumulative gain
}
\description{
Gain curve: cumulative sum of true tau (normalized to 1) when ordered by pred
}
\examples{
\dontrun{
# Basic usage
# gain_curve(...)
}
}
\seealso{
\code{\link{RCausalML-package}}
}
