% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DMLearner.R
\name{intercept.LinearDML}
\alias{intercept.LinearDML}
\alias{intercept.SparseLinearDML}
\title{Intercept of the linear CATE model (LinearDML or SparseLinearDML)}
\usage{
\method{intercept}{LinearDML}(object, ...)

\method{intercept}{SparseLinearDML}(object, ...)
}
\value{
Object returned by \code{intercept.LinearDML}.

Object returned by \code{intercept.SparseLinearDML}.
}
\description{
Intercept of the linear CATE model (LinearDML or SparseLinearDML)
}
\examples{
\dontrun{
# Basic usage
# intercept.LinearDML(...)
}
\dontrun{
# Basic usage
# intercept.SparseLinearDML(...)
}
}
\seealso{
\code{\link{RCausalML-package}}

\code{\link{RCausalML-package}}
}
