% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_learners.R
\name{intercept}
\alias{intercept}
\title{Intercept of the linear CATE model (generic)}
\usage{
intercept(object, ...)
}
\arguments{
\item{object}{fitted model (e.g. LinearDRLearner, SparseLinearDRLearner)}

\item{...}{passed to methods}
}
\value{
Object returned by \code{intercept}.
}
\description{
Intercept of the linear CATE model (generic)
}
\examples{
\dontrun{
# Basic usage
# intercept(...)
}
}
\seealso{
\code{\link{RCausalML-package}}
}
