% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cate_interpreter.R
\name{interpret}
\alias{interpret}
\title{Interpret a CATE or policy model with a tree (generic)}
\usage{
interpret(obj, cate_estimator, X, ...)
}
\arguments{
\item{obj}{interpreter object (e.g. SingleTreeCateInterpreter, SingleTreePolicyInterpreter)}

\item{cate_estimator}{fitted CATE estimator}

\item{X}{feature matrix}

\item{...}{passed to methods}
}
\value{
Object returned by \code{interpret}.
}
\description{
Interpret a CATE or policy model with a tree (generic)
}
\examples{
\dontrun{
# Basic usage
# interpret(...)
}
}
\seealso{
\code{\link{RCausalML-package}}
}
