% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notears.R
\name{is_dag}
\alias{is_dag}
\title{Check if weighted adjacency matrix W defines a DAG}
\usage{
is_dag(W)
}
\arguments{
\item{W}{Square numeric matrix (weighted adjacency).}
}
\value{
Logical.
}
\description{
Check if weighted adjacency matrix W defines a DAG
}
\examples{
\dontrun{
# Basic usage
# is_dag(...)
}
}
\seealso{
\code{\link{RCausalML-package}}
}
