% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DMLearner.R
\name{KernelDML}
\alias{KernelDML}
\title{KernelDML: DML with random Fourier features then linear CATE (EconML \code{KernelDML})}
\usage{
KernelDML(
  model_y = "ranger",
  model_t = "ranger",
  dim = 20L,
  bw = 1,
  control_name = 0,
  n_fold = 5L,
  ate_alpha = 0.05,
  seed = NULL
)
}
\arguments{
\item{model_y, model_t}{first-stage learners}

\item{dim}{number of random features (EconML kernel feature dimension)}

\item{bw}{RBF-style bandwidth scaling}

\item{control_name, n_fold, ate_alpha, seed}{as in \code{\link{DMLearner}}}
}
\value{
Object of class \code{c("KernelDML", "DMLearner")}
}
\description{
Matches EconML \code{_RandomFeatures} + linear \code{model_final} on \eqn{\phi(X)}; joint design still
cross-multiplies with \eqn{\tilde{T}}.
}
\examples{
\dontrun{
# Basic usage
# KernelDML(...)
}
}
\seealso{
\code{\link{RCausalML-package}}
}
