% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iv.R
\name{late_iv}
\alias{late_iv}
\title{Local Average Treatment Effect (LATE) via IV
LATE = (E[Y|Z=1] - E[Y|Z=0]) / (E[W|Z=1] - E[W|Z=0])}
\usage{
late_iv(Y, W, Z)
}
\arguments{
\item{Y}{outcome}

\item{W}{actual treatment (may differ from assignment due to noncompliance)}

\item{Z}{assignment (instrument)}
}
\value{
scalar LATE estimate
}
\description{
Local Average Treatment Effect (LATE) via IV
LATE = (E[Y|Z=1] - E[Y|Z=0]) / (E[W|Z=1] - E[W|Z=0])
}
\examples{
\dontrun{
# Basic usage
# late_iv(...)
}
}
\seealso{
\code{\link{RCausalML-package}}
}
