% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/causaldata_integration.R
\name{list_causaldata_datasets}
\alias{list_causaldata_datasets}
\title{List causaldata datasets supported for RCausalML}
\usage{
list_causaldata_datasets()
}
\value{
A data frame with columns: \code{name}, \code{description},
  \code{type} (binary_treatment, iv), \code{treatment}, \code{outcome},
  \code{covariates}, \code{textbook}.
}
\description{
Returns a table of dataset names and their role (treatment, outcome, IV)
for use with \code{\link{load_causaldata}}. All datasets require the
\pkg{causaldata} package.
}
\details{
Use this function first to discover supported dataset names, then pass a
selected name to \code{\link{load_causaldata}}.
}
\examples{
\dontrun{
library(RCausalML)
ds <- list_causaldata_datasets()
head(ds)
}
}
\references{
Huntington-Klein, N., & Barrett, M. (2024). causaldata: Example data sets
for causal inference textbooks. R package.
\url{https://github.com/NickCH-K/causaldata}.

Huntington-Klein, N. (2021). \emph{The Effect: An Introduction to Research
Design and Causality}. \url{https://theeffectbook.net}.

Cunningham, S. (2021). \emph{Causal Inference: The Mixtape}. Yale University
Press. \url{https://mixtape.scunning.com}.

Hernan, M. A., & Robins, J. M. (2020). \emph{Causal Inference: What If}.
Chapman & Hall/CRC. \url{https://www.hsph.harvard.edu/miguel-hernan/causal-inference-book}.
}
\seealso{
\code{\link{load_causaldata}}, \code{citation("causaldata")}
}
