% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/causalBoosted_IV.R
\name{load_boosted_iv}
\alias{load_boosted_iv}
\title{Load a BoostedIVForest model from disk.}
\usage{
load_boosted_iv(path)
}
\arguments{
\item{path}{File path written by \code{save_boosted_iv()}.}
}
\value{
A \code{BoostedIVForest} object.
}
\description{
Load a BoostedIVForest model from disk.
}
\examples{
\dontrun{
# Basic usage
# load_boosted_iv(...)
}
}
\seealso{
\code{\link{RCausalML-package}}
}
