% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/causalXGBoost.R
\name{load_cxgboost}
\alias{load_cxgboost}
\title{Load a saved CXGBoost model from disk.}
\usage{
load_cxgboost(path)
}
\arguments{
\item{path}{File path written by \code{$save_model()} or
\code{save_cxgboost()}.}
}
\value{
A \code{CXGBoost} R6 object.
}
\description{
Load a saved CXGBoost model from disk.
}
\examples{
\dontrun{
# Basic usage
# load_cxgboost(...)
}
}
\seealso{
\code{\link{RCausalML-package}}
}
