% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_arm_causal_boost.R
\name{load_multi_arm_causal_boost}
\alias{load_multi_arm_causal_boost}
\title{Load a saved MultiArmCausalBoost model}
\usage{
load_multi_arm_causal_boost(path)
}
\arguments{
\item{path}{RDS file path}
}
\value{
A MultiArmCausalBoost object
}
\description{
Load a saved MultiArmCausalBoost model
}
\examples{
\dontrun{
# Basic usage
# load_multi_arm_causal_boost(...)
}
}
\seealso{
\code{\link{RCausalML-package}}
}
