% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/causalDeepNet.R
\name{loss_function}
\alias{loss_function}
\title{CausalVAE loss function}
\usage{
loss_function(
  x,
  dec_mu,
  dec_logvar,
  enc_mu,
  enc_logvar,
  model,
  gamma_scale = 1
)
}
\arguments{
\item{x}{Input tensor.}

\item{dec_mu}{Decoder mean.}

\item{dec_logvar}{Decoder log variance.}

\item{enc_mu}{Encoder mean.}

\item{enc_logvar}{Encoder log variance.}

\item{model}{CausalVAE model instance.}

\item{gamma_scale}{Warmup scaling for structural loss term.}
}
\value{
Scalar tensor loss.
}
\description{
CausalVAE loss function
}
\examples{
\dontrun{
# Basic usage
# loss_function(...)
}
}
\seealso{
\code{\link{RCausalML-package}}
}
