% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_learners.R
\name{LRSRegressor}
\alias{LRSRegressor}
\title{LRSRegressor: S-Learner with linear regression (like Python LRSRegressor)}
\usage{
LRSRegressor(control_name = 0)
}
\value{
Object returned by \code{LRSRegressor}.
}
\description{
LRSRegressor: S-Learner with linear regression (like Python LRSRegressor)
}
\examples{
\dontrun{
# Basic usage
# LRSRegressor(...)
}
}
\seealso{
\code{\link{RCausalML-package}}
}
