% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grf_integration.R
\name{multi_arm_causal_forest}
\alias{multi_arm_causal_forest}
\title{Multi-arm causal forest (grf)}
\usage{
multi_arm_causal_forest(X, Y, W, ...)
}
\arguments{
\item{X}{The covariates.}

\item{Y}{The outcome (vector or matrix for multiple outcomes).}

\item{W}{The treatment assignment (factor).}

\item{...}{All other arguments passed to \code{\link[grf]{multi_arm_causal_forest}}.}
}
\value{
A trained \code{multi_arm_causal_forest} object.
}
\description{
Re-export of \code{\link[grf]{multi_arm_causal_forest}} from the \pkg{grf}
package. Trains a causal forest for multiple treatment arms. Requires the
\pkg{grf} package.
}
\examples{
\dontrun{
# Basic usage
# multi_arm_causal_forest(...)
}
}
\references{
Athey, S., Tibshirani, J., & Wager, S. (2019). Generalized random forests.
\emph{Annals of Statistics}, 47(2), 1148-1178.

Nie, X., & Wager, S. (2021). Quasi-oracle estimation of heterogeneous
treatment effects. \emph{Biometrika}, 108(2), 299-319.
}
\seealso{
\code{\link[grf]{multi_arm_causal_forest}}, \code{\link{grf_average_treatment_effect}}
}
