% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drawing.R
\name{multipartite_layout}
\alias{multipartite_layout}
\title{Multipartite (layered) layout}
\usage{
multipartite_layout(
  G,
  subset_key = "subset",
  align = c("vertical", "horizontal"),
  scale = 1,
  center = NULL,
  as_list = FALSE
)
}
\arguments{
\item{G}{igraph object or adjacency matrix.}

\item{subset_key}{Either character (node attribute name giving layer) or named list mapping layer id -> node ids.}

\item{align}{"vertical" or "horizontal".}

\item{scale}{Scale factor.}

\item{center}{Center coordinates.}
}
\value{
Matrix of positions (n x 2) or list if \code{as_list=TRUE}.
}
\description{
Multipartite (layered) layout
}
\examples{
\dontrun{
# Basic usage
# multipartite_layout(...)
}
}
\seealso{
\code{\link{RCausalML-package}}
}
