% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/causalDeepNet.R
\name{neuralGrangerML}
\alias{neuralGrangerML}
\title{Alias for \code{neural_granger_ml}}
\usage{
neuralGrangerML(data, ...)
}
\arguments{
\item{data}{Numeric matrix/data.frame with rows as time points and columns as variables.}
}
\value{
Object of class \code{neural_granger_ml}.
}
\description{
Alias for \code{neural_granger_ml}
}
\examples{
\dontrun{
# Basic usage
# neuralGrangerML(...)
}
}
\seealso{
\code{\link{RCausalML-package}}
}
