% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensitivity.R
\name{one_sided}
\alias{one_sided}
\title{One-sided confounding function (ATE)}
\usage{
one_sided(alpha, p, treatment)
}
\arguments{
\item{alpha}{scalar or vector; confounding strength}

\item{p}{propensity score vector (0-1)}

\item{treatment}{treatment vector (1 = treated, 0 = control)}
}
\value{
adjustment vector of length length(p)
}
\description{
One-sided confounding function (ATE)
}
\examples{
\dontrun{
# Basic usage
# one_sided(...)
}
}
\seealso{
\code{\link{RCausalML-package}}
}
