% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensitivity.R
\name{one_sided_att}
\alias{one_sided_att}
\title{One-sided confounding function for ATT}
\usage{
one_sided_att(alpha, p, treatment)
}
\arguments{
\item{alpha}{scalar or vector}

\item{p}{propensity score vector}

\item{treatment}{treatment vector}
}
\value{
adjustment vector
}
\description{
One-sided confounding function for ATT
}
\examples{
\dontrun{
# Basic usage
# one_sided_att(...)
}
}
\seealso{
\code{\link{RCausalML-package}}
}
