% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/automated_ml.R
\name{predict.AutomatedMLModel}
\alias{predict.AutomatedMLModel}
\title{Predict using fitted AutomatedMLModel}
\usage{
\method{predict}{AutomatedMLModel}(object, newdata, ...)
}
\arguments{
\item{object}{fitted \code{AutomatedMLModel}.}

\item{newdata}{matrix or data.frame of features.}

\item{...}{unused.}
}
\value{
Numeric vector of predictions, or matrix for multi-output.
}
\description{
Predict using fitted AutomatedMLModel
}
\examples{
\dontrun{
# Basic usage
# predict.AutomatedMLModel(...)
}
}
\seealso{
\code{\link{RCausalML-package}}
}
