% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/causalDeepNet.R
\name{predict.castle}
\alias{predict.castle}
\title{Predict outcome from CASTLE}
\usage{
\method{predict}{castle}(object, newdata, ...)
}
\arguments{
\item{object}{Fitted \code{castle} object.}

\item{newdata}{Matrix or data.frame containing predictors \code{X}.}

\item{...}{Ignored.}
}
\value{
Numeric vector of predicted outcomes.
}
\description{
Predict outcome from CASTLE
}
\examples{
\dontrun{
# Basic usage
# predict.castle(...)
}
}
\seealso{
\code{\link{RCausalML-package}}
}
