% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uplift_trees.R
\name{predict.causal_inference_tree}
\alias{predict.causal_inference_tree}
\title{Predict method for causal_inference_tree}
\usage{
\method{predict}{causal_inference_tree}(object, newdata, ...)
}
\value{
Object returned by \code{predict.causal_inference_tree}.
}
\description{
Predict method for causal_inference_tree
}
\examples{
\dontrun{
# Basic usage
# predict.causal_inference_tree(...)
}
}
\seealso{
\code{\link{RCausalML-package}}
}
