% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/causalDeepNet.R
\name{predict.causal_vae}
\alias{predict.causal_vae}
\title{Predict latent codes from CausalVAE}
\usage{
\method{predict}{causal_vae}(object, newdata, ...)
}
\arguments{
\item{object}{Fitted \code{causal_vae} object.}

\item{newdata}{Covariate matrix or data.frame.}

\item{...}{Ignored.}
}
\value{
Matrix of latent codes.
}
\description{
Returns latent representation \eqn{z} from a fitted \code{causal_vae} model.
}
\examples{
\dontrun{
# Basic usage
# predict.causal_vae(...)
}
}
\seealso{
\code{\link{RCausalML-package}}
}
