% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/causalDeepNet.R
\name{predict.cfrnet}
\alias{predict.cfrnet}
\title{Predict CATE from CFRNet}
\usage{
\method{predict}{cfrnet}(object, newdata, ...)
}
\arguments{
\item{object}{fitted \code{cfrnet} object}

\item{newdata}{covariate matrix or data.frame}

\item{...}{ignored}
}
\value{
Vector of predicted ITE.
}
\description{
Predict CATE from CFRNet
}
\examples{
\dontrun{
# Basic usage
# predict.cfrnet(...)
}
}
\seealso{
\code{\link{RCausalML-package}}
}
