% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DMLearner.R
\name{predict.DMLearner}
\alias{predict.DMLearner}
\title{Predict CATE from fitted DMLearner}
\usage{
\method{predict}{DMLearner}(
  object,
  newdata,
  treatment = NULL,
  y = NULL,
  return_components = FALSE,
  verbose = TRUE,
  ...
)
}
\value{
Object returned by \code{predict.DMLearner}.
}
\description{
Predict CATE from fitted DMLearner
}
\examples{
\dontrun{
# Basic usage
# predict.DMLearner(...)
}
}
\seealso{
\code{\link{RCausalML-package}}
}
