% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DMLearner.R
\name{predict.DMLIVLearner}
\alias{predict.DMLIVLearner}
\title{Predict CATE from fitted DMLIVLearner}
\usage{
\method{predict}{DMLIVLearner}(object, newdata = NULL, ...)
}
\value{
Object returned by \code{predict.DMLIVLearner}.
}
\description{
Predict CATE from fitted DMLIVLearner
}
\examples{
\dontrun{
# Basic usage
# predict.DMLIVLearner(...)
}
}
\seealso{
\code{\link{RCausalML-package}}
}
