% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_learners.R
\name{predict.DomainAdaptationLearner}
\alias{predict.DomainAdaptationLearner}
\title{Predict CATE with a fitted DomainAdaptationLearner}
\usage{
\method{predict}{DomainAdaptationLearner}(
  object,
  newdata,
  treatment = NULL,
  y = NULL,
  return_components = FALSE,
  verbose = TRUE,
  ...
)
}
\value{
Object returned by \code{predict.DomainAdaptationLearner}.
}
\description{
Predict CATE with a fitted DomainAdaptationLearner
}
\examples{
\dontrun{
# Basic usage
# predict.DomainAdaptationLearner(...)
}
}
\seealso{
\code{\link{RCausalML-package}}
}
