% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iv.R
\name{predict.driv_learner}
\alias{predict.driv_learner}
\title{Predict conditional LATE from DRIV learner}
\usage{
\method{predict}{driv_learner}(object, newdata, ...)
}
\value{
Object returned by \code{predict.driv_learner}.
}
\description{
Predict conditional LATE from DRIV learner
}
\examples{
\dontrun{
# Basic usage
# predict.driv_learner(...)
}
}
\seealso{
\code{\link{RCausalML-package}}
}
