% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_learners.R
\name{predict.DRLearner}
\alias{predict.DRLearner}
\title{Predict CATE from DR-Learner}
\usage{
\method{predict}{DRLearner}(
  object,
  newdata,
  treatment = NULL,
  y = NULL,
  return_components = FALSE,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{object}{fitted DRLearner}

\item{newdata}{feature matrix}

\item{treatment}{optional; if provided with \code{y}, used for regression metrics when \code{verbose=TRUE}}

\item{y}{optional; if provided with \code{treatment}, used for regression metrics when \code{verbose=TRUE}}

\item{return_components}{if TRUE return list with \code{te}, \code{yhat_cs}, \code{yhat_ts}}

\item{verbose}{if TRUE and \code{treatment} and \code{y} provided, print regression metrics per group}

\item{...}{unused}
}
\value{
Object returned by \code{predict.DRLearner}.
}
\description{
Predict CATE from DR-Learner
}
\examples{
\dontrun{
# Basic usage
# predict.DRLearner(...)
}
}
\seealso{
\code{\link{RCausalML-package}}
}
