% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DMLearner.R
\name{predict.DynamicDMLearner}
\alias{predict.DynamicDMLearner}
\title{Predict CATE (constant marginal effect) from fitted DynamicDMLearner}
\usage{
\method{predict}{DynamicDMLearner}(object, newdata = NULL, ...)
}
\arguments{
\item{newdata}{covariate matrix X (m x d_x); can be NULL for intercept-only}
}
\value{
Object returned by \code{predict.DynamicDMLearner}.
}
\description{
Predict CATE (constant marginal effect) from fitted DynamicDMLearner
}
\examples{
\dontrun{
# Basic usage
# predict.DynamicDMLearner(...)
}
}
\seealso{
\code{\link{RCausalML-package}}
}
