% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/causalDeepNet.R
\name{predict.ganite}
\alias{predict.ganite}
\title{Predict ITE from GANITE}
\usage{
\method{predict}{ganite}(object, newdata, ...)
}
\arguments{
\item{object}{Fitted \code{ganite} object.}

\item{newdata}{Covariate matrix or data.frame.}

\item{...}{Ignored.}
}
\value{
Vector of predicted ITE.
}
\description{
For torch GANITE, uses the Inference network to predict \eqn{\hat{Y}(0)} and \eqn{\hat{Y}(1)}, then ITE = \eqn{\hat{Y}(1) - \hat{Y}(0)}.
}
\examples{
\dontrun{
# Basic usage
# predict.ganite(...)
}
}
\seealso{
\code{\link{RCausalML-package}}
}
