% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uplift_trees.R
\name{predict.interaction_tree}
\alias{predict.interaction_tree}
\title{Predict method for interaction_tree}
\usage{
\method{predict}{interaction_tree}(object, newdata, ...)
}
\value{
Object returned by \code{predict.interaction_tree}.
}
\description{
Predict method for interaction_tree
}
\examples{
\dontrun{
# Basic usage
# predict.interaction_tree(...)
}
}
\seealso{
\code{\link{RCausalML-package}}
}
