% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/causalDeepNet.R
\name{predict.ivae}
\alias{predict.ivae}
\title{Predict latent representation from iVAE}
\usage{
\method{predict}{ivae}(object, newdata, u, ...)
}
\arguments{
\item{object}{Fitted \code{ivae} object.}

\item{newdata}{Covariate matrix or data.frame.}

\item{u}{Auxiliary labels in \code{[0, n_aux - 1]}.}

\item{...}{Ignored.}
}
\value{
Matrix of latent means.
}
\description{
Predict latent representation from iVAE
}
\examples{
\dontrun{
# Basic usage
# predict.ivae(...)
}
}
\seealso{
\code{\link{RCausalML-package}}
}
