% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DMLearner.R
\name{predict.NonParamDMLIVLearner}
\alias{predict.NonParamDMLIVLearner}
\title{Predict CATE from fitted NonParamDMLIVLearner}
\usage{
\method{predict}{NonParamDMLIVLearner}(object, newdata = NULL, ...)
}
\value{
Object returned by \code{predict.NonParamDMLIVLearner}.
}
\description{
Predict CATE from fitted NonParamDMLIVLearner
}
\examples{
\dontrun{
# Basic usage
# predict.NonParamDMLIVLearner(...)
}
}
\seealso{
\code{\link{RCausalML-package}}
}
