% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DMLearner.R
\name{predict.OrthoIVLearner}
\alias{predict.OrthoIVLearner}
\title{Predict CATE from fitted OrthoIVLearner}
\usage{
\method{predict}{OrthoIVLearner}(object, newdata = NULL, ...)
}
\value{
Object returned by \code{predict.OrthoIVLearner}.
}
\description{
Predict CATE from fitted OrthoIVLearner
}
\examples{
\dontrun{
# Basic usage
# predict.OrthoIVLearner(...)
}
}
\seealso{
\code{\link{RCausalML-package}}
}
