% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/policy_learner.R
\name{predict.policy_learner}
\alias{predict.policy_learner}
\title{Predict treatment assignment (0 or 1)}
\usage{
\method{predict}{policy_learner}(object, newdata, ...)
}
\arguments{
\item{object}{fitted \code{policy_learner} object.}

\item{newdata}{feature matrix or data.frame.}

\item{...}{unused.}
}
\value{
Integer vector of predicted treatment assignment (0 = control, 1 = treat).
}
\description{
Predict treatment assignment (0 or 1)
}
\examples{
\dontrun{
# Basic usage
# predict.policy_learner(...)
}
}
\seealso{
\code{\link{RCausalML-package}}
}
