% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cate_interpreter.R
\name{predict.SingleTreePolicyInterpreter}
\alias{predict.SingleTreePolicyInterpreter}
\title{Predict recommended treatment from SingleTreePolicyInterpreter (alias for treat)}
\usage{
\method{predict}{SingleTreePolicyInterpreter}(object, newdata, ...)
}
\arguments{
\item{object}{fitted \code{SingleTreePolicyInterpreter}.}

\item{newdata}{feature matrix or data.frame.}

\item{...}{Not used.}
}
\value{
Integer vector of recommended treatment (0 = control, 1 = first treatment, ...).
}
\description{
Predict recommended treatment from SingleTreePolicyInterpreter (alias for treat)
}
\examples{
\dontrun{
# Basic usage
# predict.SingleTreePolicyInterpreter(...)
}
}
\seealso{
\code{\link{RCausalML-package}}
}
