% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_learners.R
\name{predict.SLearner}
\alias{predict.SLearner}
\title{Predict CATE from S-Learner}
\usage{
\method{predict}{SLearner}(
  object,
  newdata,
  treatment = NULL,
  y = NULL,
  return_components = FALSE,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{object}{fitted SLearner}

\item{newdata}{covariate matrix}

\item{treatment}{optional treatment vector (for metrics if y also provided)}

\item{y}{optional outcome (for metrics if treatment provided)}

\item{return_components}{if TRUE return list(te, yhat_cs, yhat_ts)}
}
\value{
Object returned by \code{predict.SLearner}.
}
\description{
Predict CATE from S-Learner
}
\examples{
\dontrun{
# Basic usage
# predict.SLearner(...)
}
}
\seealso{
\code{\link{RCausalML-package}}
}
