% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/causalDeepNet.R
\name{predict.tarnet}
\alias{predict.tarnet}
\title{Predict CATE from TARNet}
\usage{
\method{predict}{tarnet}(object, newdata, ...)
}
\arguments{
\item{object}{fitted \code{tarnet} object}

\item{newdata}{covariate matrix or data.frame}

\item{...}{ignored}
}
\value{
Vector of predicted ITE.
}
\description{
Predict CATE from TARNet
}
\examples{
\dontrun{
# Basic usage
# predict.tarnet(...)
}
}
\seealso{
\code{\link{RCausalML-package}}
}
