% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_learners.R
\name{predict.TLearner}
\alias{predict.TLearner}
\title{Predict CATE from T-Learner}
\usage{
\method{predict}{TLearner}(
  object,
  newdata,
  treatment = NULL,
  y = NULL,
  return_components = FALSE,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{return_components}{if TRUE return list(te, yhat_cs, yhat_ts)}
}
\value{
Object returned by \code{predict.TLearner}.
}
\description{
Predict CATE from T-Learner
}
\examples{
\dontrun{
# Basic usage
# predict.TLearner(...)
}
}
\seealso{
\code{\link{RCausalML-package}}
}
