% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uplift_trees.R
\name{predict.uplift_rf_multi}
\alias{predict.uplift_rf_multi}
\title{Predict from uplift_rf_multi. full_output=TRUE returns Python-style DataFrame.}
\usage{
\method{predict}{uplift_rf_multi}(object, newdata, full_output = FALSE, ...)
}
\value{
Object returned by \code{predict.uplift_rf_multi}.
}
\description{
Predict from uplift_rf_multi. full_output=TRUE returns Python-style DataFrame.
}
\examples{
\dontrun{
# Basic usage
# predict.uplift_rf_multi(...)
}
}
\seealso{
\code{\link{RCausalML-package}}
}
