% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uplift_trees.R
\name{predict.uplift_tree_ddp}
\alias{predict.uplift_tree_ddp}
\title{Predict from single uplift tree (DDP, IDDP, IT, CIT)}
\usage{
\method{predict}{uplift_tree_ddp}(object, newdata, ...)
}
\value{
Object returned by \code{predict.uplift_tree_ddp}.
}
\description{
Predict from single uplift tree (DDP, IDDP, IT, CIT)
}
\examples{
\dontrun{
# Basic usage
# predict.uplift_tree_ddp(...)
}
}
\seealso{
\code{\link{RCausalML-package}}
}
