% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uplift_trees.R
\name{predict.uplift_tree_iddp}
\alias{predict.uplift_tree_iddp}
\title{Predict method for uplift_tree_iddp}
\usage{
\method{predict}{uplift_tree_iddp}(object, newdata, ...)
}
\value{
Object returned by \code{predict.uplift_tree_iddp}.
}
\description{
Predict method for uplift_tree_iddp
}
\examples{
\dontrun{
# Basic usage
# predict.uplift_tree_iddp(...)
}
}
\seealso{
\code{\link{RCausalML-package}}
}
