% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimize.R
\name{predict_best_treatment.counterfactual_value_estimator}
\alias{predict_best_treatment.counterfactual_value_estimator}
\title{Predict best treatment from value estimator}
\usage{
\method{predict_best_treatment}{counterfactual_value_estimator}(obj, ...)
}
\value{
Object returned by \code{predict_best_treatment.counterfactual_value_estimator}.
}
\description{
Predict best treatment from value estimator
}
\examples{
\dontrun{
# Basic usage
# predict_best_treatment.counterfactual_value_estimator(...)
}
}
\seealso{
\code{\link{RCausalML-package}}
}
