% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimize.R
\name{predict_best_treatment}
\alias{predict_best_treatment}
\title{Predict best treatment for each unit}
\usage{
predict_best_treatment(obj, ...)
}
\value{
Object returned by \code{predict_best_treatment}.
}
\description{
Predict best treatment for each unit
}
\examples{
\dontrun{
# Basic usage
# predict_best_treatment(...)
}
}
\seealso{
\code{\link{RCausalML-package}}
}
