% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/automated_ml.R
\name{predict_proba.AutomatedMLModel}
\alias{predict_proba.AutomatedMLModel}
\title{Predict class probabilities using fitted AutomatedMLModel (classification only)}
\usage{
\method{predict_proba}{AutomatedMLModel}(object, newdata, ...)
}
\arguments{
\item{object}{fitted \code{AutomatedMLModel} with \code{task = "classification"}.}

\item{newdata}{matrix or data.frame of features.}

\item{...}{unused.}
}
\value{
Matrix of probabilities (columns = classes).
}
\description{
Predict class probabilities using fitted AutomatedMLModel (classification only)
}
\examples{
\dontrun{
# Basic usage
# predict_proba.AutomatedMLModel(...)
}
}
\seealso{
\code{\link{RCausalML-package}}
}
