% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/policy_learner.R
\name{predict_proba.DRPolicyForest}
\alias{predict_proba.DRPolicyForest}
\title{Predict probability of recommending each treatment (DR Policy Forest)
Average of tree probabilities.}
\usage{
\method{predict_proba}{DRPolicyForest}(object, newdata, ...)
}
\arguments{
\item{object}{fitted \code{DRPolicyForest}.}

\item{newdata}{feature matrix or data.frame.}

\item{...}{unused.}
}
\value{
Matrix (n_samples x n_treatments).
}
\description{
Predict probability of recommending each treatment (DR Policy Forest)
Average of tree probabilities.
}
\examples{
\dontrun{
# Basic usage
# predict_proba.DRPolicyForest(...)
}
}
\seealso{
\code{\link{RCausalML-package}}
}
