% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/policy_learner.R
\name{predict_proba.policy_learner}
\alias{predict_proba.policy_learner}
\title{Predict probability of treatment assignment (P(treat))}
\usage{
\method{predict_proba}{policy_learner}(object, newdata, ...)
}
\arguments{
\item{object}{fitted \code{policy_learner} object.}

\item{newdata}{feature matrix or data.frame.}

\item{...}{unused.}
}
\value{
Numeric vector of predicted probability of recommending treatment.
}
\description{
Predict probability of treatment assignment (P(treat))
}
\examples{
\dontrun{
# Basic usage
# predict_proba.policy_learner(...)
}
}
\seealso{
\code{\link{RCausalML-package}}
}
