% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/policy_learner.R
\name{predict_proba}
\alias{predict_proba}
\title{Predict probability of treatment assignment (P(treat))}
\usage{
predict_proba(object, newdata, ...)
}
\arguments{
\item{object}{fitted model object.}

\item{newdata}{feature matrix or data.frame.}

\item{...}{passed to methods.}
}
\value{
Object returned by \code{predict_proba}.
}
\description{
Generic for predicted probability of recommending treatment.
}
\examples{
\dontrun{
# Basic usage
# predict_proba(...)
}
}
\seealso{
\code{\link{RCausalML-package}}
}
