% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/policy_learner.R
\name{predict_value}
\alias{predict_value}
\title{Predict expected value under each treatment (policy value)}
\usage{
predict_value(object, newdata, ...)
}
\arguments{
\item{object}{fitted policy model.}

\item{newdata}{feature matrix or data.frame.}

\item{...}{passed to methods.}
}
\value{
Object returned by \code{predict_value}.
}
\description{
Predict expected value under each treatment (policy value)
}
\examples{
\dontrun{
# Basic usage
# predict_value(...)
}
}
\seealso{
\code{\link{RCausalML-package}}
}
