% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/causalForest.R
\name{print.causal_forest}
\alias{print.causal_forest}
\title{Print method for causal_forest}
\usage{
print.causal_forest(x, ...)
}
\value{
Object returned by \code{print.causal_forest}.
}
\description{
Print method for causal_forest
}
\examples{
\dontrun{
# Basic usage
# print.causal_forest(...)
}
}
\seealso{
\code{\link{RCausalML-package}}
}
