% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/propensity.R
\name{propensity_glm}
\alias{propensity_glm}
\title{Logistic regression propensity score}
\usage{
propensity_glm(X, w, ...)
}
\arguments{
\item{X}{covariate matrix or data.frame (will add intercept)}

\item{w}{treatment indicator}

\item{...}{passed to \code{glm}}
}
\value{
vector of propensity scores
}
\description{
Simple GLM propensity model. For elastic net + calibration use
\code{propensity_model_logistic_regression()} or \code{compute_propensity_score()}.
}
\examples{
\dontrun{
# Basic usage
# propensity_glm(...)
}
}
\seealso{
\code{\link{RCausalML-package}}
}
