% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{qini_curve}
\alias{qini_curve}
\title{Qini curve: cumulative true treatment effect when units ordered by predicted CATE (descending)}
\usage{
qini_curve(tau, pred)
}
\arguments{
\item{tau}{true treatment effect vector}

\item{pred}{predicted CATE vector}
}
\value{
cumulative sum of tau in order of pred (descending)
}
\description{
Qini curve: cumulative true treatment effect when units ordered by predicted CATE (descending)
}
\examples{
\dontrun{
# Basic usage
# qini_curve(...)
}
}
\seealso{
\code{\link{RCausalML-package}}
}
