% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{qini_score_vec}
\alias{qini_score_vec}
\title{Qini score: area under Qini curve (trapezoidal)}
\usage{
qini_score_vec(tau, pred)
}
\arguments{
\item{tau}{true treatment effect vector}

\item{pred}{predicted CATE vector}
}
\value{
scalar
}
\description{
Qini score: area under Qini curve (trapezoidal)
}
\examples{
\dontrun{
# Basic usage
# qini_score_vec(...)
}
}
\seealso{
\code{\link{RCausalML-package}}
}
