% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grf_integration.R
\name{rank_average_treatment_effect.fit}
\alias{rank_average_treatment_effect.fit}
\title{Rank-weighted average treatment effect fit (grf)}
\usage{
rank_average_treatment_effect.fit(DR.scores, priorities, ...)
}
\arguments{
\item{DR.scores}{Doubly robust evaluation scores.}

\item{priorities}{Treatment prioritization scores.}

\item{...}{All other arguments passed to \code{\link[grf]{rank_average_treatment_effect.fit}}.}
}
\value{
An object of class \code{rank_average_treatment_effect}.
}
\description{
Re-export of \code{\link[grf]{rank_average_treatment_effect.fit}} from the
\pkg{grf} package. Computes RATE with user-supplied doubly robust scores.
Requires the \pkg{grf} package.
}
\examples{
\dontrun{
# Basic usage
# rank_average_treatment_effect.fit(...)
}
}
\references{
Yadlowsky, S., Fleming, S., Shah, N., Brunskill, E., & Wager, S. (2025).
Evaluating treatment prioritization rules via rank-weighted average
treatment effects. \emph{JASA}, 120(549), 1-12.
}
\seealso{
\code{\link[grf]{rank_average_treatment_effect.fit}}, \code{\link{rank_average_treatment_effect}}
}
