\name{rcausalml-extra-objects}
\alias{rcausalml-extra-objects}
\alias{MultiArmCausalBoost}
\alias{multi_arm_ATE}
\alias{multi_arm_PEHE}
\alias{coef}
\title{Additional exported RCausalML objects}
\description{
Documentation placeholder for exported objects that are also documented in
other function/class help pages.
}
\details{
\code{MultiArmCausalBoost} is an R6 class for multi-arm causal boosting.
\code{multi_arm_ATE} and \code{multi_arm_PEHE} are evaluation helpers used by that class.
\code{coef} is the exported S3 generic with methods documented on their respective
class help pages.
}
\seealso{
\code{\link{load_multi_arm_causal_boost}},
\code{\link{save_multi_arm_causal_boost}},
\code{\link{run_multi_arm_causal_boost_example}}
}
