% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{regression_metrics}
\alias{regression_metrics}
\title{Regression metrics (MSE, MAE, R2) by treatment group}
\usage{
regression_metrics(y, yhat, w)
}
\arguments{
\item{y}{observed outcome}

\item{yhat}{predicted outcome}

\item{w}{treatment indicator (0/1)}
}
\value{
list with mse, mae, r2 for control and treated
}
\description{
Regression metrics (MSE, MAE, R2) by treatment group
}
\examples{
\dontrun{
# Basic usage
# regression_metrics(...)
}
}
\seealso{
\code{\link{RCausalML-package}}
}
