% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/causalXGBoost.R
\name{save_cxgboost}
\alias{save_cxgboost}
\title{Save a CXGBoost model to disk (standalone wrapper).}
\usage{
save_cxgboost(model, path)
}
\arguments{
\item{model}{A fitted \code{CXGBoost} object.}

\item{path}{File path.}
}
\value{
Object returned by \code{save_cxgboost}.
}
\description{
Save a CXGBoost model to disk (standalone wrapper).
}
\examples{
\dontrun{
# Basic usage
# save_cxgboost(...)
}
}
\seealso{
\code{\link{RCausalML-package}}
}
