% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_arm_causal_boost.R
\name{save_multi_arm_causal_boost}
\alias{save_multi_arm_causal_boost}
\title{Save a MultiArmCausalBoost model}
\usage{
save_multi_arm_causal_boost(model, path)
}
\arguments{
\item{model}{A fitted MultiArmCausalBoost object}

\item{path}{File path}
}
\value{
Object returned by \code{save_multi_arm_causal_boost}.
}
\description{
Save a MultiArmCausalBoost model
}
\examples{
\dontrun{
# Basic usage
# save_multi_arm_causal_boost(...)
}
}
\seealso{
\code{\link{RCausalML-package}}
}
