% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notears.R
\name{set_random_seed}
\alias{set_random_seed}
\title{Set random seed for reproducibility (R and torch)}
\usage{
set_random_seed(seed)
}
\arguments{
\item{seed}{Integer seed.}
}
\value{
Object returned by \code{set_random_seed}.
}
\description{
Set random seed for reproducibility (R and torch)
}
\examples{
\dontrun{
# Basic usage
# set_random_seed(...)
}
}
\seealso{
\code{\link{RCausalML-package}}
}
