% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drawing.R
\name{shell_layout}
\alias{shell_layout}
\title{Shell layout: place nodes in concentric circles (shells)}
\usage{
shell_layout(
  G,
  nlist = NULL,
  rotate = NULL,
  scale = 1,
  center = NULL,
  as_list = FALSE
)
}
\arguments{
\item{G}{igraph object or adjacency matrix.}

\item{nlist}{List of integer vectors; each vector gives node indices (1-based) in that shell. If NULL, all nodes in one shell.}

\item{rotate}{Angle in radians to rotate each shell (default pi/n_shells).}

\item{scale}{Scale factor.}

\item{center}{Center coordinates.}
}
\value{
Matrix of positions (n x 2) or list if \code{as_list=TRUE}.
}
\description{
Shell layout: place nodes in concentric circles (shells)
}
\examples{
\dontrun{
# Basic usage
# shell_layout(...)
}
}
\seealso{
\code{\link{RCausalML-package}}
}
