% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notears.R
\name{simulate_dag}
\alias{simulate_dag}
\title{Simulate random DAG with expected number of edges}
\usage{
simulate_dag(d, s0, graph_type = "ER")
}
\arguments{
\item{d}{Number of nodes.}

\item{s0}{Expected number of edges.}

\item{graph_type}{"ER" (Erdos-Renyi), "SF" (scale-free).}
}
\value{
Binary adjacency matrix (d x d).
}
\description{
Simulate random DAG with expected number of edges
}
\examples{
\dontrun{
# Basic usage
# simulate_dag(...)
}
}
\seealso{
\code{\link{RCausalML-package}}
}
