% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{simulate_easy_propensity_difficult_baseline}
\alias{simulate_easy_propensity_difficult_baseline}
\title{Synthetic data: easy propensity, difficult baseline (Setup C)}
\usage{
simulate_easy_propensity_difficult_baseline(
  n = 1000,
  p = 5,
  sigma = 1,
  adj = 0
)
}
\value{
Object returned by \code{simulate_easy_propensity_difficult_baseline}.
}
\description{
Synthetic data: easy propensity, difficult baseline (Setup C)
}
\examples{
\dontrun{
# Basic usage
# simulate_easy_propensity_difficult_baseline(...)
}
}
\seealso{
\code{\link{RCausalML-package}}
}
