% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{simulate_hidden_confounder}
\alias{simulate_hidden_confounder}
\title{Synthetic data: hidden confounder (Louizos et al.)}
\usage{
simulate_hidden_confounder(n = 10000, p = 5, sigma = 1, adj = 0)
}
\value{
Object returned by \code{simulate_hidden_confounder}.
}
\description{
Synthetic data: hidden confounder (Louizos et al.)
}
\examples{
\dontrun{
# Basic usage
# simulate_hidden_confounder(...)
}
}
\seealso{
\code{\link{RCausalML-package}}
}
