% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notears.R
\name{simulate_linear_sem}
\alias{simulate_linear_sem}
\title{Simulate samples from linear SEM}
\usage{
simulate_linear_sem(W, n, sem_type = "gauss", noise_scale = NULL)
}
\arguments{
\item{W}{Weighted adjacency matrix (d x d) of a DAG.}

\item{n}{Number of samples.}

\item{sem_type}{"gauss", "exp", "uniform".}

\item{noise_scale}{Scalar or length-d vector; default 1.}
}
\value{
Sample matrix (n x d).
}
\description{
Simulate samples from linear SEM
}
\examples{
\dontrun{
# Basic usage
# simulate_linear_sem(...)
}
}
\seealso{
\code{\link{RCausalML-package}}
}
