% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{simulate_nuisance_and_easy_treatment}
\alias{simulate_nuisance_and_easy_treatment}
\title{Synthetic data with difficult nuisance and easy treatment (Setup A)}
\usage{
simulate_nuisance_and_easy_treatment(n = 1000, p = 5, sigma = 1, adj = 0)
}
\arguments{
\item{n}{number of observations}

\item{p}{number of covariates (>= 5)}

\item{sigma}{standard deviation of error}

\item{adj}{propensity adjustment (higher -> propensity shifted toward 0)}
}
\value{
list with y, X, w, tau, b, e
}
\description{
Synthetic data with difficult nuisance and easy treatment (Setup A)
}
\examples{
\dontrun{
# Basic usage
# simulate_nuisance_and_easy_treatment(...)
}
}
\seealso{
\code{\link{RCausalML-package}}
}
