% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notears.R
\name{simulate_parameter}
\alias{simulate_parameter}
\title{Simulate SEM edge weights for a DAG}
\usage{
simulate_parameter(B, w_ranges = list(c(-2, -0.5), c(0.5, 2)))
}
\arguments{
\item{B}{Binary adjacency matrix (d x d).}

\item{w_ranges}{List of (low, high) ranges for weights.}
}
\value{
Weighted adjacency matrix (d x d).
}
\description{
Simulate SEM edge weights for a DAG
}
\examples{
\dontrun{
# Basic usage
# simulate_parameter(...)
}
}
\seealso{
\code{\link{RCausalML-package}}
}
