% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{simulate_randomized_trial}
\alias{simulate_randomized_trial}
\title{Synthetic data: randomized trial (Setup B)}
\usage{
simulate_randomized_trial(n = 1000, p = 5, sigma = 1, adj = 0)
}
\value{
Object returned by \code{simulate_randomized_trial}.
}
\description{
Synthetic data: randomized trial (Setup B)
}
\examples{
\dontrun{
# Basic usage
# simulate_randomized_trial(...)
}
}
\seealso{
\code{\link{RCausalML-package}}
}
