% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{simulate_unrelated_treatment_control}
\alias{simulate_unrelated_treatment_control}
\title{Synthetic data: unrelated treatment and control (Setup D)}
\usage{
simulate_unrelated_treatment_control(n = 1000, p = 5, sigma = 1, adj = 0)
}
\value{
Object returned by \code{simulate_unrelated_treatment_control}.
}
\description{
Synthetic data: unrelated treatment and control (Setup D)
}
\examples{
\dontrun{
# Basic usage
# simulate_unrelated_treatment_control(...)
}
}
\seealso{
\code{\link{RCausalML-package}}
}
