% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match.R
\name{smd}
\alias{smd}
\title{Standardized Mean Difference (SMD) of a feature between treatment and control}
\usage{
smd(feature, treatment)
}
\arguments{
\item{feature}{numeric vector (one covariate)}

\item{treatment}{binary vector (1 = treatment, 0 = control)}
}
\value{
numeric SMD value
}
\description{
Definition: (mean_t - mean_c) / sqrt(0.5 * (var_t + var_c)).
See https://www.ncbi.nlm.nih.gov/pmc/articles/PMC3144483/#s11title
}
\examples{
\dontrun{
# Basic usage
# smd(...)
}
}
\seealso{
\code{\link{RCausalML-package}}
}
