% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drawing.R
\name{spiral_layout}
\alias{spiral_layout}
\title{Spiral layout}
\usage{
spiral_layout(
  G,
  scale = 1,
  center = NULL,
  resolution = 0.35,
  equidistant = FALSE,
  as_list = FALSE
)
}
\arguments{
\item{G}{igraph object or adjacency matrix.}

\item{scale}{Scale factor.}

\item{center}{Center coordinates.}

\item{resolution}{Compactness (lower = more compressed).}

\item{equidistant}{If TRUE, equidistant nodes along spiral; else equal angles.}
}
\value{
Matrix of positions (n x 2) or list if \code{as_list=TRUE}.
}
\description{
Spiral layout
}
\examples{
\dontrun{
# Basic usage
# spiral_layout(...)
}
}
\seealso{
\code{\link{RCausalML-package}}
}
