% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/causalForest.R
\name{summary.causal_forest}
\alias{summary.causal_forest}
\title{Summary method for causal_forest}
\usage{
summary.causal_forest(object, ...)
}
\value{
Object returned by \code{summary.causal_forest}.
}
\description{
Summary method for causal_forest
}
\examples{
\dontrun{
# Basic usage
# summary.causal_forest(...)
}
}
\seealso{
\code{\link{RCausalML-package}}
}
