% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cate_interpreter.R
\name{treat}
\alias{treat}
\title{Assign treatment from a policy interpreter (generic)}
\usage{
treat(obj, X, ...)
}
\arguments{
\item{obj}{fitted SingleTreePolicyInterpreter}

\item{X}{feature matrix}

\item{...}{passed to methods}
}
\value{
Object returned by \code{treat}.
}
\description{
Assign treatment from a policy interpreter (generic)
}
\examples{
\dontrun{
# Basic usage
# treat(...)
}
}
\seealso{
\code{\link{RCausalML-package}}
}
