% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/causalDeepNet.R
\name{tune_hyperparameters}
\alias{tune_hyperparameters}
\title{Tune CausalVAE hyperparameters}
\usage{
tune_hyperparameters(
  train_loader,
  val_loader,
  n_trials = 10L,
  n_epochs_per_trial = 50L,
  input_dim = 3L,
  latent_dim = 3L,
  device = NULL
)
}
\arguments{
\item{train_loader}{Training dataloader.}

\item{val_loader}{Validation dataloader.}

\item{n_trials}{Number of random trials.}

\item{n_epochs_per_trial}{Number of epochs per trial.}

\item{input_dim}{Input dimension.}

\item{latent_dim}{Latent dimension.}

\item{device}{Device string.}
}
\value{
List of best parameters.
}
\description{
Tune CausalVAE hyperparameters
}
\examples{
\dontrun{
# Basic usage
# tune_hyperparameters(...)
}
}
\seealso{
\code{\link{RCausalML-package}}
}
