% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uplift_trees.R
\name{uplift_rf_chi}
\alias{uplift_rf_chi}
\title{Uplift RF with Chi-squared criterion}
\usage{
uplift_rf_chi(
  X,
  treatment,
  y,
  n_trees = 100,
  min_node_size = 10,
  min_samples_treatment = 10,
  max_depth = 5,
  max_features = NULL,
  n_reg = 100,
  normalization = TRUE,
  random_state = NULL,
  n_cores = 1L,
  ...
)
}
\arguments{
\item{n_cores}{number of cores for building trees (1 = sequential)}
}
\value{
Object returned by \code{uplift_rf_chi}.
}
\description{
Uplift RF with Chi-squared criterion
}
\examples{
\dontrun{
# Basic usage
# uplift_rf_chi(...)
}
}
\seealso{
\code{\link{RCausalML-package}}
}
