% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uplift_trees.R
\name{uplift_tree_ddp}
\alias{uplift_tree_ddp}
\title{Uplift tree on delta-delta-p (DDP) criterion; binary outcome only}
\usage{
uplift_tree_ddp(X, treatment, y, min_node_size = 10, max_depth = 5)
}
\value{
Object returned by \code{uplift_tree_ddp}.
}
\description{
Uplift tree on delta-delta-p (DDP) criterion; binary outcome only
}
\examples{
\dontrun{
# Basic usage
# uplift_tree_ddp(...)
}
}
\seealso{
\code{\link{RCausalML-package}}
}
