% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uplift_trees.R
\name{uplift_tree_iddp}
\alias{uplift_tree_iddp}
\title{Uplift tree on IDDP criterion; binary outcome only}
\usage{
uplift_tree_iddp(X, treatment, y, min_node_size = 10, max_depth = 5)
}
\value{
Object returned by \code{uplift_tree_iddp}.
}
\description{
Uplift tree on IDDP criterion; binary outcome only
}
\examples{
\dontrun{
# Basic usage
# uplift_tree_iddp(...)
}
}
\seealso{
\code{\link{RCausalML-package}}
}
