% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uplift_trees.R
\name{uplift_tree_to_rpart}
\alias{uplift_tree_to_rpart}
\title{Convert an uplift tree (list) to an rpart object for use with rpart.plot}
\usage{
uplift_tree_to_rpart(tree, x_names = NULL)
}
\arguments{
\item{tree}{tree list (from \code{build_uplift_tree} or \code{uplift_rf_kl(...)$trees[[1]]})}

\item{x_names}{character vector of feature names}
}
\value{
An object of class \code{rpart} that can be passed to \code{rpart.plot()},
  or \code{NULL} if \pkg{rpart} is not available.
}
\description{
Builds a minimal \code{rpart} object so that \code{rpart.plot::rpart.plot()} can
be used. Useful when you want the polished layout and options of the
\pkg{rpart.plot} package (see \url{http://www.milbo.org/rpart-plot/index.html}).
}
\examples{
\dontrun{
# Basic usage
# uplift_tree_to_rpart(...)
}
}
\seealso{
\code{\link{RCausalML-package}}
}
