% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/causalForest.R
\name{variable_importance}
\alias{variable_importance}
\title{Variable importance from split frequency}
\usage{
variable_importance(forest)
}
\arguments{
\item{forest}{A causal_forest object}
}
\value{
Named numeric vector of importance scores (sums to 1)
}
\description{
Variable importance from split frequency
}
\examples{
\dontrun{
# Basic usage
# variable_importance(...)
}
}
\seealso{
\code{\link{RCausalML-package}}
}
