% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_learners.R
\name{XGBRRegressor}
\alias{XGBRRegressor}
\title{XGBRRegressor: R-Learner with XGBoost}
\usage{
XGBRRegressor(control_name = 0, n_fold = 5)
}
\arguments{
\item{control_name}{value for control group (0 or "control")}

\item{n_fold}{number of folds for propensity (if p not provided)}
}
\value{
Object returned by \code{XGBRRegressor}.
}
\description{
XGBRRegressor: R-Learner with XGBoost
}
\examples{
\dontrun{
# Basic usage
# XGBRRegressor(...)
}
}
\seealso{
\code{\link{RCausalML-package}}
}
