% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_learners.R
\name{XGBSRegressor}
\alias{XGBSRegressor}
\title{XGBSRegressor: S-Learner with XGBoost}
\usage{
XGBSRegressor(control_name = 0)
}
\arguments{
\item{control_name}{value for control group (0 or "control")}
}
\value{
Object returned by \code{XGBSRegressor}.
}
\description{
XGBSRegressor: S-Learner with XGBoost
}
\examples{
\dontrun{
# Basic usage
# XGBSRegressor(...)
}
}
\seealso{
\code{\link{RCausalML-package}}
}
