% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_learners.R
\name{XGBTRegressor}
\alias{XGBTRegressor}
\title{XGBTRegressor: T-Learner with XGBoost (like Python XGBTRegressor)}
\usage{
XGBTRegressor(control_name = 0)
}
\arguments{
\item{control_name}{value for control group (0 or "control")}
}
\value{
Object returned by \code{XGBTRegressor}.
}
\description{
XGBTRegressor: T-Learner with XGBoost (like Python XGBTRegressor)
}
\examples{
\dontrun{
# Basic usage
# XGBTRegressor(...)
}
}
\seealso{
\code{\link{RCausalML-package}}
}
