% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_learners.R
\name{XGBXRegressor}
\alias{XGBXRegressor}
\title{XGBXRegressor: X-Learner with XGBoost}
\usage{
XGBXRegressor(control_name = 0)
}
\arguments{
\item{control_name}{value for control group (0 or "control")}
}
\value{
Object returned by \code{XGBXRegressor}.
}
\description{
XGBXRegressor: X-Learner with XGBoost
}
\examples{
\dontrun{
# Basic usage
# XGBXRegressor(...)
}
}
\seealso{
\code{\link{RCausalML-package}}
}
